add followWikiLink()
This one is complicated. HTMLElement.click() is only supposed to work
on clickable elements.

At least on chrome when that method is called on a <wiki-link> it
causes an error when the click handler tries to call preventDefault()
on the click event object. It does end up adding the page to the
lineup.

I found the dispatchEvent(new MouseEvent("click")) method. It works
nicely for clicking <wiki-link> elements. No error raised about
preventDefault(). But when clicking collaborative links in a
node-based wiki, it seems to miss the normal wiki handler. Instead of
adding the new page to the lineup, it appears to do a full page load
of only the new page.